Update CA.
The US Supreme Court has ruled on Hollingworth v Perry: it is remanded
to the 9th Circuit with instruction to dismiss for lack of standing.
Thus (once the dust settles) the ruling in Perry v Schwarzenegger
stands, Prop 8 is in violation of the US Constitution, and California
has equal marriage.